+++ atropos/__init__.py
@@ -2,7 +2,7 @@
 from __future__ import print_function, division, absolute_import
 import sys
 
-__version__ = '1.0.15'
+__version__ = '1.0.16'
 
 def check_importability():  # pragma: no cover
     try: